+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 * Fix most of #294, thanks @haf
 * Fix #285 for LiterateConsole. This should be it. Thanks @haf.
 * Move colourising tokenisers to a module more easily discoverable. Thanks @haf
+* Fix #300 naming of Events.miniLevel -> minLevel
 
 #### 5.0.0-beta.3
 * Targets: AliYun target ported to .Net Core